Design data working now Resolves #28
