+++ source/GitPocket.Business.Test/Package.appxmanifest
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Package xmlns="http://schemas.microsoft.com/appx/2010/manifest">
   
-  <Identity Name="989d5039-b76d-46be-950b-f9a5fc63a0b7"
+  <Identity Name="c8b1aea8-086d-4927-b586-51dad75e749f"
             Publisher="CN=Alberto"
             Version="1.0.0.0" />
