+++ source/GitPocket.Business/packages.config
@@ -1,8 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="BugSense.XAML81" version="3.6.6.0" targetFramework="portable-win81+wpa81" />
-  <package id="CommonServiceLocator" version="1.3" targetFramework="portable-win81+wpa81" />
-  <package id="MvvmLightLibs" version="5.0.2.0" targetFramework="portable-win81+wpa81" />
-  <package id="Newtonsoft.Json" version="6.0.8" targetFramework="portable-win81+wpa81" />
   <package id="Octokit" version="0.6.2" targetFramework="portable-win81+wpa81" />
 </packages>
