+++ source/GitPocket.Test/GitPocket.Test.csproj
@@ -6,7 +6,7 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProductVersion>8.0.30703</ProductVersion>
     <SchemaVersion>2.0</SchemaVersion>
-    <ProjectGuid>{D1C7FF2C-C6A4-4357-80AB-A674C6743362}</ProjectGuid>
+    <ProjectGuid>{D31044CA-7ACB-4104-9999-68FDEF7EDB0B}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>GitPocket.Test</RootNamespace>
