+++ source/GitPocket.Test/Package.appxmanifest
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Package xmlns="http://schemas.microsoft.com/appx/2010/manifest">
   
-  <Identity Name="5d6d9de4-c4fe-4a40-a0a6-7036bc21aa94"
+  <Identity Name="adebb388-c009-4182-892b-e5952a3f4387"
             Publisher="CN=Alberto"
             Version="1.0.0.0" />
