+++ source/GitPocket.Test/UnitTest1.cs
@@ -1,4 +1,8 @@
-﻿using Microsoft.VisualStudio.TestPlatform.UnitTestFramework;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using Microsoft.VisualStudio.TestPlatform.UnitTestFramework;
 
 namespace GitPocket.Test
 {
