+++ source/GitPocket/GitPocket.Shared/App.xaml.cs
