--- source/GitPocket/GitPocket.Shared/IoC/INinjectServiceLocator.cs
@@ -1,10 +1,0 @@
-using Microsoft.Practices.ServiceLocation;
-using Ninject.Syntax;
-
-namespace GitPocket.IoC
-{
-    public interface INinjectServiceLocator : IServiceLocator
-    {
-        IBindingToSyntax<T> Bind<T>();
-    }
-}
