+++ source/GitPocket/GitPocket.Shared/MainPage.xaml
@@ -1,16 +1,15 @@
-﻿<themeManagerRt:ThemeEnabledPage xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-      xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
-      xmlns:Core="using:Microsoft.Xaml.Interactions.Core"
-      xmlns:themeManagerRt="using:ThemeManagerRt"
-      x:Class="GitPocket.MainPage"
-      mc:Ignorable="d"
-      Background="{ThemeResource ApplicationPageBackgroundThemeBrush}"
-      Loaded="Page_Loaded" DataContext="{Binding Main, Mode=OneWay, Source={StaticResource Locator}}">
-    <Grid>
+﻿<Page
+    x:Class="GitPocket.MainPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:GitPocket"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:i="using:Microsoft.Xaml.Interactivity"
+    xmlns:core="using:Microsoft.Xaml.Interactions.Core"
+    mc:Ignorable="d" DataContext="{Binding Main, Mode=OneWay, Source={StaticResource Locator}}">
 
+    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
         <Grid.RowDefinitions>
             <RowDefinition Height="Auto" />
             <RowDefinition Height="Auto" />
@@ -61,21 +60,21 @@
                     <Storyboard>
                         <DoubleAnimation Duration="0" To="1" Storyboard.TargetProperty="(UIElement.Opacity)" Storyboard.TargetName="grid1" d:IsOptimized="True"/>
                         <ObjectAnimationUsingKeyFrames Storyboard.TargetProperty="(TextBlock.FontFamily)" Storyboard.TargetName="textBlock">
-                        	<DiscreteObjectKeyFrame KeyTime="0">
-                        		<DiscreteObjectKeyFrame.Value>
-                        			<FontFamily>Global User Interface</FontFamily>
-                        		</DiscreteObjectKeyFrame.Value>
-                        	</DiscreteObjectKeyFrame>
+                            <DiscreteObjectKeyFrame KeyTime="0">
+                                <DiscreteObjectKeyFrame.Value>
+                                    <FontFamily>Global User Interface</FontFamily>
+                                </DiscreteObjectKeyFrame.Value>
+                            </DiscreteObjectKeyFrame>
                         </ObjectAnimationUsingKeyFrames>
                     </Storyboard>
                 </VisualState>
             </VisualStateGroup>
         </VisualStateManager.VisualStateGroups>
-        <interactivity:Interaction.Behaviors>
-            <Core:EventTriggerBehavior EventName="Loaded">
-                <Core:GoToStateAction StateName="Init" />
-            </Core:EventTriggerBehavior>
-        </interactivity:Interaction.Behaviors>
+        <i:Interaction.Behaviors>
+            <core:EventTriggerBehavior EventName="Loaded">
+                <core:GoToStateAction StateName="Init" />
+            </core:EventTriggerBehavior>
+        </i:Interaction.Behaviors>
         <TextBlock x:Uid="SignInText"
                    Name="SignInTextBlock"
                    TextWrapping="Wrap"
@@ -111,12 +110,12 @@
                 Grid.Row="3"
                 VerticalAlignment="Top" Height="59" Width="339" Background="{StaticResource ButtonBackgroundBrush}" BorderBrush="{x:Null}"
                 BorderThickness="0" Margin="0,15,0,0" Opacity="0">
-            <interactivity:Interaction.Behaviors>
-                <Core:EventTriggerBehavior EventName="Click">
-                    <Core:GoToStateAction StateName="LoginIn" />
-                    <Core:InvokeCommandAction Command="{Binding OnLogin}"/>
-                </Core:EventTriggerBehavior>
-            </interactivity:Interaction.Behaviors>
+            <i:Interaction.Behaviors>
+                <core:EventTriggerBehavior EventName="Click">
+                    <core:GoToStateAction StateName="LoginIn" />
+                    <core:InvokeCommandAction Command="{Binding OnLogin}"/>
+                </core:EventTriggerBehavior>
+            </i:Interaction.Behaviors>
         </Button>
         <Grid x:Name="grid" Opacity="0" Grid.RowSpan="5" Background="Black" IsHitTestVisible="False">
             <StackPanel Orientation="Vertical" HorizontalAlignment="Center" VerticalAlignment="Center">
@@ -164,4 +163,4 @@
             </StackPanel>
         </Grid>
     </Grid>
-</themeManagerRt:ThemeEnabledPage>
+</Page>
