+++ source/GitPocket/GitPocket.Shared/ViewModel/MainViewModel.cs
@@ -1,9 +1,9 @@
-﻿using System;
+using System;
 using System.Threading.Tasks;
-using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Media.Imaging;
 using GalaSoft.MvvmLight;
 using GalaSoft.MvvmLight.Command;
+using Windows.UI.Xaml.Media;
 using GitPocket.Business;
 
 namespace GitPocket.ViewModel
