+++ source/GitPocket/GitPocket.Windows/Properties/AssemblyInfo.cs
