+++ source/GitPocket/GitPocket.Windows/Properties/DesignTimeResources.xaml
@@ -1,0 +1,9 @@
+﻿<ResourceDictionary
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" 
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:GitPocket">
+	<ResourceDictionary.MergedDictionaries>
+		<ResourceDictionary Source="ms-appx:///Style/DefaultTheme.xaml"/>
+	</ResourceDictionary.MergedDictionaries>
+    
+</ResourceDictionary>
