+++ source/GitPocket/GitPocket.Windows/packages.config
@@ -3,7 +3,6 @@
   <package id="CommonServiceLocator" version="1.3" targetFramework="win81" />
   <package id="MvvmLight" version="5.0.2.0" targetFramework="win81" />
   <package id="MvvmLightLibs" version="5.0.2.0" targetFramework="win81" />
-  <package id="Newtonsoft.Json" version="6.0.8" targetFramework="win81" />
   <package id="Octokit" version="0.6.2" targetFramework="win81" />
-  <package id="Portable.Ninject" version="3.3.1" targetFramework="win81" />
+  <package id="ThemeManagerRt" version="1.0.0.1" targetFramework="win81" />
 </packages>
