+++ source/GitPocket/GitPocket.WindowsPhone/Package.appxmanifest
@@ -1,29 +1,41 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Package xmlns="http://schemas.microsoft.com/appx/2010/manifest" xmlns:m2="http://schemas.microsoft.com/appx/2013/manifest" xmlns:m3="http://schemas.microsoft.com/appx/2014/manifest" xmlns:mp="http://schemas.microsoft.com/appx/2014/phone/manifest">
-  <Identity Name="e1b09090-b163-4bfe-a24b-11635dc2063c" Publisher="CN=Alberto Monteiro" Version="1.0.0.0" />
-  <mp:PhoneIdentity PhoneProductId="e1b09090-b163-4bfe-a24b-11635dc2063c" PhonePublisherId="00000000-0000-0000-0000-000000000000" />
+
+  <Identity Name="f00ddfb4-bd81-4d26-86be-9270cd45392d"
+            Publisher="CN=Alberto"
+            Version="1.0.0.0" />
+
+  <mp:PhoneIdentity PhoneProductId="f00ddfb4-bd81-4d26-86be-9270cd45392d" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
+
   <Properties>
     <DisplayName>GitPocket.WindowsPhone</DisplayName>
-    <PublisherDisplayName>Alberto Monteiro</PublisherDisplayName>
+    <PublisherDisplayName>Alberto</PublisherDisplayName>
     <Logo>Assets\StoreLogo.png</Logo>
   </Properties>
+
   <Prerequisites>
     <OSMinVersion>6.3.1</OSMinVersion>
     <OSMaxVersionTested>6.3.1</OSMaxVersionTested>
   </Prerequisites>
+  
   <Resources>
-    <Resource Language="x-generate" />
+    <Resource Language="x-generate"/>
   </Resources>
+
   <Applications>
-    <Application Id="App" Executable="$targetnametoken$.exe" EntryPoint="GitPocket.WindowsPhone.App">
-      <m3:VisualElements DisplayName="GitPocket" Square150x150Logo="Assets\Logo.png" Square44x44Logo="Assets\SmallLogo.png" Description="GitPocket" ForegroundText="light" BackgroundColor="transparent">
-        <m3:DefaultTile Wide310x150Logo="Assets\WideLogo.png" Square71x71Logo="Assets\Square71x71Logo.png">
-          <m3:ShowNameOnTiles>
-            <m3:ShowOn Tile="square150x150Logo" />
-          </m3:ShowNameOnTiles>
-        </m3:DefaultTile>
-        <m3:SplashScreen Image="Assets\SplashScreen.png" />
-      </m3:VisualElements>
+    <Application Id="App"
+        Executable="$targetnametoken$.exe"
+        EntryPoint="GitPocket.WindowsPhone.App">
+        <m3:VisualElements
+            DisplayName="GitPocket.WindowsPhone"
+            Square150x150Logo="Assets\Logo.png"
+            Square44x44Logo="Assets\SmallLogo.png"
+            Description="GitPocket.WindowsPhone"
+            ForegroundText="light"
+            BackgroundColor="transparent">
+            <m3:DefaultTile Wide310x150Logo="Assets\WideLogo.png" Square71x71Logo="Assets\Square71x71Logo.png"/>
+            <m3:SplashScreen Image="Assets\SplashScreen.png"/>
+        </m3:VisualElements>
     </Application>
   </Applications>
   <Capabilities>
