--- source/GitPocket/GitPocket.WindowsPhone/Properties/DesignTimeResources.xaml
@@ -1,8 +1,0 @@
-﻿<ResourceDictionary
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" 
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:GitPocket">
-	<ResourceDictionary.MergedDictionaries>
-		<ResourceDictionary Source="ms-appx:///Style/DefaultTheme.xaml"/>
-	</ResourceDictionary.MergedDictionaries>
-</ResourceDictionary>
