--- source/GitPocket/GitPocket.WindowsPhone/Views/MainPage.xaml.cs
@@ -1,55 +1,0 @@
-using System;
-using Windows.UI.ViewManagement;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Navigation;
-using GitPocket.ViewModel;
-
-namespace GitPocket
-{
-    public sealed partial class MainPage
-    {
-        public MainPage()
-        {
-            InitializeComponent();
-
-            NavigationCacheMode = NavigationCacheMode.Required;
-        }
-
-        /// <summary>
-        ///     Invoked when this page is about to be displayed in a Frame.
-        /// </summary>
-        /// <param name="e">
-        ///     Event data that describes how this page was reached.
-        ///     This parameter is typically used to configure the page.
-        /// </param>
-        protected override void OnNavigatedTo(NavigationEventArgs e)
-        {
-            // TODO: Prepare page for display here.
-
-            // TODO: If your application contains multiple pages, ensure that you are
-            // handling the hardware Back button by registering for the
-            // Windows.Phone.UI.Input.HardwareButtons.BackPressed event.
-            // If you are using the NavigationHelper provided by some templates,
-            // this event is handled for you.
-        }
-
-        private async void Page_Loaded(object sender, Windows.UI.Xaml.RoutedEventArgs e)
-        {
-#if WINDOWS_PHONE_APP
-            StatusBar statusBar = StatusBar.GetForCurrentView();
-
-            await statusBar.HideAsync(); 
-#endif
-            var mainViewModel = (MainViewModel)DataContext;
-
-            mainViewModel.PropertyChanged += (o, p) =>
-            {
-                if (p.PropertyName == "State")
-                {
-                    var viewModel = (MainViewModel)o;
-                    VisualStateManager.GoToState(this, viewModel.State, true);
-                }
-            };
-        }
-    }
-}
