+++ source/GitPocket/GitPocket.WindowsPhone/packages.config
@@ -6,6 +6,5 @@
   <package id="MvvmLightLibs" version="5.0.2.0" targetFramework="wpa81" />
   <package id="Newtonsoft.Json" version="6.0.8" targetFramework="wpa81" />
   <package id="Octokit" version="0.6.2" targetFramework="wpa81" />
-  <package id="Portable.Ninject" version="3.3.1" targetFramework="wpa81" />
   <package id="ThemeManagerRt" version="1.0.0.1" targetFramework="wpa81" />
 </packages>
